+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'dart:math';
 import 'dart:ui';
 
 import 'package:cached_network_image/cached_network_image.dart';
@@ -17,15 +18,44 @@ class MainApp extends StatefulWidget {
 
 class _MainAppState extends State<MainApp> {
   bool enable = true;
+  final text = 'This is a spoiler! Tap to reveal';
 
   final url =
       'https://img.freepik.com/premium-photo/drawing-female-superhero-female-character_1308175-151081.jpg?w=1800';
 
+  SpoilerMask createStarPath(Size size) {
+    Path path = Path();
+    final cx = size.width / 2;
+    final cy = size.height / 3;
+    final radiusOuter = size.width;
+    final radiusInner = radiusOuter / 2;
+    const numPoints = 5;
+
+    const angle = pi / (numPoints);
+    path.moveTo(cx + radiusOuter * cos(0), cy + radiusOuter * sin(0));
+
+    for (int i = 1; i <= numPoints * 2; i++) {
+      final r = (i % 2 == 0) ? radiusOuter : radiusInner;
+      final x = cx + r * cos(i * angle);
+      final y = cy + r * sin(i * angle);
+      path.lineTo(x, y);
+    }
+    path.close();
+
+    const rotationAngle = pi / 3;
+
+    // Rotate the path using a transformation matrix
+    final Matrix4 matrix = Matrix4.identity()
+      ..translate(cx + 280, cy)
+      ..rotateZ(rotationAngle)
+      ..translate(-cx, -cy);
+    return SpoilerMask(maskPath: path.transform(matrix.storage), maskOperation: PathOperation.union);
+  }
+
   @override
   Widget build(BuildContext context) {
-    const text = 'This is a spoiler! Tap to reveal';
-
     return MaterialApp(
+      // showPerformanceOverlay: true,
       home: Scaffold(
         backgroundColor: Colors.black,
         body: Column(
@@ -33,35 +63,36 @@ class _MainAppState extends State<MainApp> {
           children: [
             RepaintBoundary(
               child: SpoilerText(
-                config: TextSpoilerConfiguration(
+                config: TextSpoilerConfig(
                   isEnabled: enable,
                   maxParticleSize: 1,
                   particleColor: Colors.white,
                   particleDensity: .1,
-                  speedOfParticles: .2,
+                  particleSpeed: .2,
                   fadeRadius: 3,
-                  fadeAnimation: true,
-                  enableGesture: true,
-                  // selection: const TextSelection(baseOffset: 0, extentOffset: 19),
-                  style: const TextStyle(
+                  enableFadeAnimation: true,
+                  enableGestureReveal: true,
+                  // textSelection: const TextSelection(baseOffset: 0, extentOffset: 12),
+                  textStyle: const TextStyle(
                     fontSize: 50,
                     color: Colors.white,
                   ),
+                  maskConfig: createStarPath(const Size.square(80)),
                 ),
                 text: text,
               ),
             ),
             RepaintBoundary(
               child: SpoilerOverlay(
-                config: WidgetSpoilerConfiguration(
+                config: WidgetSpoilerConfig(
                   isEnabled: enable,
                   maxParticleSize: 1,
                   particleDensity: .1,
-                  speedOfParticles: .4,
+                  particleSpeed: .4,
                   particleColor: Colors.white,
                   fadeRadius: 3,
-                  fadeAnimation: true,
-                  enableGesture: true,
+                  enableFadeAnimation: true,
+                  enableGestureReveal: true,
                   imageFilter: ImageFilter.blur(sigmaX: 30.0, sigmaY: 30.0),
                 ),
                 child: CachedNetworkImage(imageUrl: url),